# imobiliary-api/openapi.yaml
openapi: 3.1.0
info:
  title: Imobiliary API
  version: "0.9.0"
  summary: Rental property management for brokers, small agencies and owners.
  description: |
    The API behind the Imobiliary platform. It covers identity and the office
    (accounts, sessions with a second factor, members and invitations, the
    account holder's rights under the LGPD) and the office's register of
    people and properties. Contracts and rents arrive in later phases.

    ## Offices and roles

    Data belongs to an office (an organisation). An account works inside one
    office at a time and may belong to several. Inside an office it is either
    `admin`, who manages members, invitations and the office itself, or
    `member`. An office is never left without an administrator.

    ## Authentication

    A session is two credentials.

    - The **access token** is a JWT signed with Ed25519, valid for 15 minutes,
      sent as `Authorization: Bearer <token>`. It names the account and the
      office. The membership behind it is read again on every request, so a
      role change or a removal takes effect at once, not when the token
      expires. A token minted before the account's last password change is
      refused.
    - The **refresh token** is an opaque secret exchanged at
      `POST /v1/sessions/refresh` for a new pair. Each one is spent once.
      Presenting a spent one ends every session of the account, because two
...

    ## The second factor

...

...

    ## Errors

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearer: []
paths:
...
  /v1/accounts:
    post:
      tags: [Accounts]
      summary: Open an account and its office
      description: |
        Creates the account and the office it administers, with the account as
        its administrator. The first sign-in will ask for second-factor
        enrolment.

        Answers `202` with no body for any well-formed request, including an
        address that already has an account. The owner of that address is told
        by mail instead. Nobody is signed in by this call.
      operationId: register
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name, password, organization_name, terms_version]
              properties:
                email: { type: string, format: email, maxLength: 254 }
                name: { type: string, minLength: 1, maxLength: 120 }
                password: { $ref: "#/components/schemas/NewPassword" }
                organization_name: { type: string, minLength: 1, maxLength: 120 }
                terms_version:
                  type: string
                  description: The version of the terms the person accepted.
                  examples: ["1.0"]
      responses:
        "202": { description: "Accepted. Check the mailbox, then sign in." }
        "400": { $ref: "#/components/responses/BadRequest" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/password/forgot:
    post:
      tags: [Accounts]
      summary: Ask for a reset link
      description: |
        Mails a link to set a new password, valid for 30 minutes and usable
        once. Always `202`, whether or not the address has an account.
      operationId: forgotPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string }
      responses:
        "202": { description: Accepted. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/password/reset:
    post:
      tags: [Accounts]
      summary: Set a new password from the link
      description: |
        Spends the link and replaces the password. Every session of the account
        ends, and any other reset link stops working.
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string, description: The secret from the link. }
                password: { $ref: "#/components/schemas/NewPassword" }
      responses:
        "204": { description: The password was set. Sign in with it. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }

  # --- sessions ---------------------------------------------------------------
  /v1/sessions:
    post:
      tags: [Sessions]
      summary: Sign in
      description: |
        Checks the password. The answer is either a session, or, for an
        account with a second factor, a challenge to complete at
        `POST /v1/sessions/mfa`. Both list the offices the account belongs to.

        Without `organization_id` the session opens in the first office by
        name. Naming an office the account does not belong to is `403`, as is
        an account that belongs to no office any more.
      operationId: signIn
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
                organization_id: { $ref: "#/components/schemas/ID" }
      responses:
        "200":
          description: A session, or the challenge for the second factor.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SignInResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }

    delete:
      tags: [Sessions]
      summary: Sign out
      description: |
        Revokes the refresh token. Always `204`, whether or not it matched a
        live session. The access token in hand keeps working until it expires,
        at most 15 minutes; a client discards it.
      operationId: signOut
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "204": { description: Signed out. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/sessions/mfa:
    post:
      tags: [Sessions]
      summary: Complete a sign-in with the second factor
      description: |
        Spends the challenge, right code or wrong. A refused code means signing
        in again for a new challenge. `code` is the six digits from the app or
        one of the recovery codes; a recovery code is spent too.
      operationId: completeSecondFactor
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challenge, code]
              properties:
                challenge: { type: string }
                code: { type: string }
                organization_id: { $ref: "#/components/schemas/ID" }
      responses:
        "200":
          description: The session.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/sessions/refresh:
    post:
      tags: [Sessions]
      summary: Exchange a refresh token for a new pair
      description: |
        Spends the refresh token and returns a new session in the same office.
        A spent token presented again ends every session of the account. Two
        concurrent requests with one token: one succeeds, the other is `401`.
        A client should share one refresh between its concurrent calls.
      operationId: refreshSession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RefreshRequest" }
      responses:
        "200":
          description: The new session.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/sessions/docgen-token:
    post:
      tags: [Sessions]
      summary: A short token for the document service
      description: |
        Mints a five-minute token for the caller's session, audience `docgen`,
        signed with the same Ed25519 key as access tokens. It carries `sub`,
        `org`, `org_name`, `email`, `name` and `role`. The document service
        verifies it with the public keys (`imobiliary public-keys` prints
        them); this API refuses it, since its audience is not this API's.
      operationId: issueDocgenToken
      responses:
        "200":
          description: The token.
          content:
            application/json:
              schema:
                type: object
                required: [token, expires_at]
                properties:
                  token: { type: string }
                  expires_at: { $ref: "#/components/schemas/Instant" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/sessions/switch:
    post:
      tags: [Sessions]
      summary: Move the session to another office
      description: |
        Spends the refresh token and returns a session in the named office,
        which the account must belong to.
      operationId: switchOrganization
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token, organization_id]
              properties:
                refresh_token: { type: string }
                organization_id: { $ref: "#/components/schemas/ID" }
      responses:
        "200":
          description: The session in the other office.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Session" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }

  # --- me ---------------------------------------------------------------------
  /v1/me:
    get:
      tags: [Me]
      summary: The signed-in account
      operationId: getMe
      responses:
        "200":
          description: The account, where it works and what it may do.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Me" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/me/password:
    post:
      tags: [Me]
      summary: Change the password
      description: |
        Replaces the password for someone who knows the current one. **Every
        session ends, this one included**: the access token that made this
        request is refused from now on, so the client signs in again. The
        account's address is told by mail.
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string }
                new_password: { $ref: "#/components/schemas/NewPassword" }
      responses:
        "204": { description: Changed. Sign in again. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }

  # --- second factor ----------------------------------------------------------
  /v1/me/totp:
    post:
      tags: [Second factor]
      summary: Start enrolment
      description: |
        Generates a secret and stores it unconfirmed. Nothing changes for
        signing in until `POST /v1/me/totp/confirm` proves a code from it.
        Starting again replaces the previous attempt; an account whose second
        factor is already on gets `409`. Reachable by a session that still owes
        enrolment.
      operationId: startEnrollment
      responses:
        "200":
          description: The secret, as base32 and as an otpauth URI for a QR code.
          content:
            application/json:
              schema:
                type: object
                required: [secret, uri]
                properties:
                  secret: { type: string, examples: ["JBSWY3DPEHPK3PXP"] }
                  uri:
                    type: string
                    description: An `otpauth://totp/` address with the issuer and the account's email in its label.
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "429": { $ref: "#/components/responses/RateLimited" }

    delete:
      tags: [Second factor]
      summary: Turn the second factor off
      description: An administrator may not; that is `403 forbidden`.
      operationId: disableSecondFactor
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordRequest" }
      responses:
        "204": { description: Turned off. The recovery codes are gone too. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/me/totp/confirm:
    post:
      tags: [Second factor]
      summary: Confirm enrolment with a code
      description: |
        Proves a code from the app and turns the second factor on. The step of
        the accepted code is recorded, so the next code must come from a later
        step. The ten recovery codes in the response exist in readable form
        only here; the service keeps their digests.

        `409` when there is no enrolment to confirm.
      operationId: confirmEnrollment
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CodeRequest" }
      responses:
        "200":
          description: On, with the recovery codes.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecoveryCodes" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/me/totp/recovery-codes:
    post:
      tags: [Second factor]
      summary: Replace the recovery codes
      description: |
        Issues ten new codes. The previous list stops working. `409` when the
        second factor is off.
      operationId: regenerateRecoveryCodes
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordRequest" }
      responses:
        "200":
          description: The new codes, shown once.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RecoveryCodes" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "429": { $ref: "#/components/responses/RateLimited" }

  # --- privacy ----------------------------------------------------------------
  /v1/me/export:
    get:
      tags: [Privacy]
      summary: A copy of the account's data
      description: |
        The access right of LGPD art. 18, II: the account, its memberships,
        the second-factor state, stored sessions, invitations received, the
        audit events the account acted in and its access records, oldest first.

        No secret is included: no password hash, TOTP secret, token or code
        digest. The export is recorded in the audit trail. Reachable by a
        session that still owes enrolment; charged to the write budget.
      operationId: exportAccount
      responses:
        "200":
          description: The copy.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountExport" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/me/deletion:
    post:
      tags: [Privacy]
      summary: Delete the account
      description: |
        The erasure right of LGPD art. 18, VI, confirmed with the password.
        Immediate and permanent. A POST with a body, rather than a DELETE on
        `/v1/me`, because proxies may drop the body of a DELETE.

        **Erased:** the account, its memberships, sessions, second factor,
        recovery codes and pending resets. An office whose only member is this
        account is erased with it.

        **Kept, as the law requires:** the access records (Marco Civil art. 15)
        keep the account's identifier, and the account's address is kept
        sealed with the field key so they still identify someone. Both are
        purged six months after the deletion (LGPD art. 16, I). The audit
        trail keeps its entries without the actor; each office the account
        belonged to gets a `user.deleted` entry.

        **Refused** with `422` on `organizations` while the account is the only
        administrator of an office that has other members. Make someone else an
        administrator first. Also refused, with `422` on `organization_data`,
        when an office whose only member is this account still holds people:
        an office's data is never erased as a side effect of closing an account.

        Reachable by a session that still owes enrolment. Charged to both the
        credential limit and the write budget. The address is mailed afterwards.
      operationId: deleteAccount
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PasswordRequest" }
      responses:
        "204": { description: Deleted. Every credential the client holds is now refused. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }

  # --- office -----------------------------------------------------------------
...
  /v1/organization/members:
    get:
      tags: [Office]
      summary: Who works in the office
      description: Every member may read it. Ordered by name.
      operationId: listMembers
      responses:
        "200":
          description: The members.
          content:
            application/json:
              schema:
                type: object
                required: [members]
                properties:
                  members:
                    type: array
                    items: { $ref: "#/components/schemas/Member" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
components:
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT (EdDSA)
  schemas:
...
    Instant:
      type: string
      format: date-time
...
    NewPassword:
      type: string
      minLength: 12
      maxLength: 256
      description: Length is the only rule. At least 12 characters.
...
  responses:
    BadRequest:
      description: The body could not be read, or is larger than 1 MiB (`413`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    UnsupportedMediaType:
      description: The body must be `application/json`.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...